requirements.txt
trimesh
opencv-python
click
pytorch3d
requests
matplotlib
numpy
pillow
tqdm
torchvision
